Module ic_nns_governance::governance

The 'governance' module contains the canister (smart contract) that manages neurons, proposals, voting, voter following, voting rewards, and the code necessary to execute accepted proposals.

The governance canister interfaces with the 'ledger' canister to deal with the transfer of ICP (Internet Computer Protocol) tokens between ledger accounts, and with the 'registry' canister to distribute configuration information to all nodes of all subnetworks that participate in the Internet Computer (IC).

Structs

Governance
The Governance canister implements the full public interface of the IC's governance system.

Constants

EXECUTE_NNS_FUNCTION_PAYLOAD_LISTING_BYTES_MAX

MAX_DISSOLVE_DELAY_SECONDS

MAX_FOLLOWEES_PER_TOPIC
The maximum amount of followees each neuron can establish for each topic.

MAX_NEURON_AGE_FOR_AGE_BONUS

MAX_NEURON_RECENT_BALLOTS
The maximum amount of recent ballots to keep, per neuron.

MAX_NUMBER_OF_OPEN_MANAGE_NEURON_PROPOSALS
The max number of open manage neuron proposals.

MAX_NUMBER_OF_PROPOSALS_WITH_BALLOTS
The max number of unsettled proposals -- that is proposals for which ballots are still stored.

MIN_DISSOLVE_DELAY_FOR_VOTE_ELIGIBILITY_SECONDS
The minimum dissolve delay so that a neuron may vote.

ONE_DAY_SECONDS

ONE_MONTH_SECONDS

ONE_YEAR_SECONDS

PROPOSAL_MOTION_TEXT_BYTES_MAX

REWARD_DISTRIBUTION_PERIOD_SECONDS
The desired period for reward distribution events.

Traits

Environment
A general trait for the environment in which governance is running.

Ledger