Struct ic_nns_governance::governance::Governance

pub struct Governance {
    pub proto: GovernanceProto,
    pub env: Box<dyn Environment>,
    pub topic_followee_index: BTreeMap<Topic, BTreeMap<u64, BTreeSet<u64>>>,
    pub principal_to_neuron_ids_index: BTreeMap<PrincipalId, HashSet<u64>>,
    pub latest_gc_timestamp_seconds: u64,
    pub latest_gc_num_proposals: usize,
    // some fields omitted
}

The Governance canister implements the full public interface of the IC's governance system.

Fields
proto: GovernanceProto

The Governance Protobuf which contains all persistent state of the IC's governance system. Needs to be stored and retrieved on upgrades.

env: Box<dyn Environment>

Implementation of Environment to make unit testing easier.

topic_followee_index: BTreeMap<Topic, BTreeMap<u64, BTreeSet<u64>>>

Cached data structure that (for each topic) maps a followee to the set of followers. This is the inverse of the mapping from neuron (follower) to followees, in the neurons. This is a cached index and will be removed and recreated when the state is saved and restored.

Topic -> (neuron ID of followee) -> set of followers.

principal_to_neuron_ids_index: BTreeMap<PrincipalId, HashSet<u64>>

Maps Principals to the Neuron IDs of all Neurons that have this Principal as their controller or as one of their hot keys

This is a cached index and will be removed and recreated when the state is saved and restored.

latest_gc_timestamp_seconds: u64

The time of the latest "garbage collection" - when obsolete proposals were cleaned up.

latest_gc_num_proposals: usize

The number of proposals after the last time GC was run.

Implementations

impl Governance

pub fn new(
proto: GovernanceProto,
env: Box<dyn Environment>,
ledger: Box<dyn Ledger>
) -> Self

pub fn validate(&self) -> Result<(), GovernanceError>

Validates that the underlying protobuf is well formed.

pub fn get_neuron(&self, id: &NeuronId) -> Result<&Neuron, GovernanceError>

pub fn get_neuron_mut(
&mut self,
id: &NeuronId
) -> Result<&mut Neuron, GovernanceError>

pub fn get_neuron_ids_by_principal(&self, principal: &PrincipalId) -> Vec<u64>

Return the Neuron IDs of all Neurons that have principal as their controller or as one of their hot keys

pub fn list_neurons_by_principal(
&self,
req: &ListNeurons,
caller: &PrincipalId
) -> ListNeuronsResponse

See ListNeurons.

pub fn get_neuron_by_subaccount(
&self,
subaccount: &Subaccount
) -> Option<&Neuron>

Returns a neuron, given a subaccount.

Currently we just do linear search on the neurons. We tried an index at some point, but the index was too big, took too long to build and ultimately lowered our max possible amount of neurons, so we "downgraded" to linear search.

Consider changing this if getting a neuron by subaccount ever gets in a hot path.

pub fn get_neuron_by_subaccount_mut(
&mut self,
subaccount: &Subaccount
) -> Option<&mut Neuron>

pub async fn create_or_refresh_neuron<'_>(
&'_ mut self,
transfer: NeuronStakeTransfer
) -> Result<NeuronId, GovernanceError>

Creates a new neuron or refreshes the stake of an existing one from a ledger transfer.

Since this is meant to be called by the ledger, we're limited in what we can do/return. We expect to be provided all the information about the transfer (can't do anything without that), but after we do have all the information the only failure mode is to return the transfer back.

Pre-conditions: In all cases: 1 - The sender of the transfer matches the principal encoded in the hash of the subaccount. That is, the sender of the transfer is the controller of the neuron. 2 - The nonce of the subaccount's hash is the memo of the transfer.

For new neurons, i.e. if the subaccount of the transfer is a new one:

The new neuron won't take us above the MAX_NUMBER_OF_NEURONS.

The amount transfered was greater than or equal to self.enconomics.neuron_minimum_stake_e8s.

The transfer's subaccount is provided and is not being used by an existing neuron.

For neuron refresh, i.e. if the subaccount of the transfer already exists.

Post-conditions:

If all the pre-conditions apply, either a new neuron is created or the stake of an existing neuron is updated.

If one of more of the pre-conditions do not apply, the stake is transferred back to the source account.

pub fn claim_gtc_neurons(
&mut self,
caller: &PrincipalId,
new_controller: PrincipalId,
neuron_ids: Vec<NeuronId>
) -> Result<(), GovernanceError>

Claim the neurons supplied by the GTC on behalf of new_controller

For each neuron ID in neuron_ids, check that the corresponding neuron exists in self.proto.neurons and the neuron's controller is the GTC. If the neuron is in the expected state, set the neuron's controller to new_controller and set other fields (e.g. created_timestamp_seconds).

pub async fn transfer_gtc_neuron<'_, '_, '_, '_>(
&'_ mut self,
caller: &'_ PrincipalId,
donor_neuron_id: &'_ NeuronId,
recipient_neuron_id: &'_ NeuronId
) -> Result<(), GovernanceError>

Transfer a GTC neuron to a recipient neuron.

This will transfer the stake of the donor neuron to the recipient neuron, and perform a ledger transfer from the donor neuron's sub-account to the recipient neuron's sub-account. The donor neuron will then be deleted.

pub async fn disburse_neuron<'_, '_, '_, '_>(
&'_ mut self,
id: &'_ NeuronId,
caller: &'_ PrincipalId,
disburse: &'_ Disburse
) -> Result<u64, GovernanceError>

Disburse the stake of a neuron.

This causes the stake of a neuron to be disbursed to the provided principal (and optional subaccount). If amount is provided then that amount is disbursed.

Note that we don't enforce that 'amount' is actually smaller than or equal to the cached stake in the neuron. This will allow a user to still disburse funds if:

Someone transferred more funds to the neuron's subaccount after the the initial neuron claim that we didn't know about.

The transfer of funds previously failed for some reason (e.g. the ledger was unavailable or broken).

On success returns the block height at which the transfer happened.

Preconditions:

The neuron exists.

The caller is the controller of the the neuron.

The neuron's state is Dissolved at the current timestamp.

pub async fn split_neuron<'_, '_, '_, '_>(
&'_ mut self,
id: &'_ NeuronId,
caller: &'_ PrincipalId,
split: &'_ Split
) -> Result<NeuronId, GovernanceError>

Splits a neuron into two neurons.

The parent neuron's stake is decreased by the amount specified in Split, while the child neuron is created with a stake equal to that amount, minus the transfer fee.

The child neuron inherits all the properties of its parent including age and dissolve state.

On success returns the newly created neuron's id.

Preconditions:

The parent neuron exists

The caller is the controller of the neuron

The parent neuron is not already undergoing ledger updates.

The staked amount minus amount to split is more than the minimum stake.

The amount to split minus the transfer fee is more than the minimum stake.

pub async fn spawn_neuron<'_, '_, '_, '_>(
&'_ mut self,
id: &'_ NeuronId,
caller: &'_ PrincipalId,
spawn: &'_ Spawn
) -> Result<NeuronId, GovernanceError>

Spawn an neuron from an existing neuron's maturity.

This spawns a new neuron from an existing neuron's maturity. The existing neuron must have enough accumulated maturity such that the new neuron has stake that is more than the minimum stake.

The newly spawned neuron has the dissolve delay specified in NetworkEconomics.

Pre-conditions:

The parent neuron exists.

The caller is the controller of the neuron.

The parent neuron is not already undergoing ledger updates.

The parent neuron has accumulated maturity that would generate more than Networkeconomics::neuron_minimum_spawn_stake_e8s staked in the child neuron.

pub async fn disburse_to_neuron<'_, '_, '_, '_>(
&'_ mut self,
id: &'_ NeuronId,
caller: &'_ PrincipalId,
disburse_to_neuron: &'_ DisburseToNeuron
) -> Result<NeuronId, GovernanceError>

Disburse part of the stake of a neuron into a new neuron, possibly owned by someone else and with a different dissolve delay.

The parent neuron's stake is decreased by the amount specified in DisburseToNeuron, while the child neuron is created with a stake equal to that amount, minus the transfer fee.

The child neuron doesn't inherit any of the properties of the parent neuron, except its following.

On success returns the newly created neuron's id.

Preconditions:

The parent neuron exists

The caller is the controller of the neuron

The parent neuron is not already undergoing ledger updates.

The parent neuron's state is Dissolved at the current timestamp.

The staked amount minus amount to split is more than the minimum stake.

The amount to split minus the transfer fee is more than the minimum stake.

pub fn set_proposal_execution_status(
&mut self,
pid: u64,
result: Result<(), GovernanceError>
)

Set the status of a proposal that is 'being executed' to 'executed' or 'failed' depending on the value of 'succcess'.

The proposal ID 'pid' is taken as a raw integer to avoid lifetime issues.

pub fn get_neuron_info(
&self,
id: &NeuronId
) -> Result<NeuronInfo, GovernanceError>

Returns the neuron info for a given neuron id. This method does not require authorization, so the NeuronInfo of a neuron is accessible to any caller.

pub fn get_full_neuron(
&self,
id: &NeuronId,
caller: &PrincipalId
) -> Result<Neuron, GovernanceError>

Returns the complete neuron data for a given neuron id after checking that the caller is authorized. The neuron's controller and hot keys are authorized, as are the controllers and hot keys of any neurons that are listed as followees of the requested neuron on the ManageNeuron topic.

pub fn get_node_providers(&self) -> &[NodeProvider]

pub fn latest_reward_event(&self) -> &RewardEvent

pub fn get_proposal_info(
&self,
caller: &PrincipalId,
pid: impl Into<ProposalId>
) -> Option<ProposalInfo>

Tries to get a proposal given a proposal id

The proposal's ballots only show votes from neurons that the caller either controls or is a registered hot key for.

pub fn get_pending_proposals(&self, caller: &PrincipalId) -> Vec<ProposalInfo>

Gets all open proposals

The proposals' ballots only show votes from neurons that the caller either controls or is a registered hot key for.

Proposals with ExecuteNnsFunction as action have their payload cleared if larger than EXECUTE_NNS_FUNCTION_PAYLOAD_LISTING_BYTES_MAX. The caller can retrieve dropped payloads by calling get_proposal_info for each proposal of interest.

pub fn get_pending_proposals_data(&self) -> impl Iterator<Item = &ProposalData>

Iterator over proposals info of pending proposals.

pub fn get_proposal_data(
&self,
pid: impl Into<ProposalId>
) -> Option<&ProposalData>

pub fn list_proposals(
&self,
caller: &PrincipalId,
req: &ListProposalInfo
) -> ListProposalInfoResponse

Returns the proposals info of proposals with proposal ID less than before_proposal (exclusive), returning at most limit proposal infos. If before_proposal is not provided, start from the highest available proposal ID (inclusive).

As proposal IDs are assigned sequentially, this retrives up to limit proposals older (in terms of creation) than a specific proposal. This can be used to paginate through proposals, as follows:

let mut lst = gov.list_proposals(ListProposalInfo {}); while !lst.empty() { /* do stuff with lst */ lst = gov.list_proposals(ListProposalInfo { before_proposal: lst.last().and_then(|x|x.id) }); }

A proposal with resticted voting is included only if the caller is allowed to vote on the proposal.

The proposals' ballots only show votes from neurons that the caller either controls or is a registered hot key for.

Proposals with ExecuteNnsFunction as action have their payload cleared if larger than EXECUTE_NNS_FUNCTION_PAYLOAD_LISTING_BYTES_MAX. The caller can retrieve dropped payloads by calling get_proposal_info for each proposal of interest.

pub fn process_proposal(&mut self, pid: u64)

This method attempts to move a proposal forward in the process, from open to adopted or rejected, to executed or failed (for a previously adopted proposal).

If the proposal is open, it tallies the ballots and updates the yes, no, and undecided voting power accordingly.

This may result in the proposal becoming adopted or rejected.

If a proposal is adopted but not executed, this method attempts to execute it.

pub fn approve_genesis_kyc(&mut self, principals: &[PrincipalId])

Mark all Neurons controlled by the given principals as having passed KYC verification

pub fn make_proposal(
&mut self,
proposer_id: &NeuronId,
caller: &PrincipalId,
proposal: &Proposal
) -> Result<ProposalId, GovernanceError>

pub async fn manage_neuron<'_, '_, '_>(
&'_ mut self,
caller: &'_ PrincipalId,
mgmt: &'_ ManageNeuron
) -> ManageNeuronResponse

pub fn maybe_gc(&mut self) -> bool

Garbage collect obsolete data from the governance canister.

Current implementation only garbage collects proposals - not neurons.

Returns true if GC was run and false otherwise.

pub async fn run_periodic_tasks<'_>(&'_ mut self)

Triggers a reward distribution event if enough time has passed since the last one. This is intended to be called by a cron process.