Trait ic_nns_governance::governance::Environment

pub trait Environment: Send + Sync {
    fn now(&self) -> u64;

    fn random_u64(&mut self) -> u64;

    fn random_byte_array(&mut self) -> [u8; 32];

    fn execute_nns_function(
        &self, 
        proposal_id: u64, 
        update: &ExecuteNnsFunction
    ) -> Result<(), GovernanceError>;
}

A general trait for the environment in which governance is running.

Required methods

fn now(&self) -> u64

Returns the current time, in seconds since the epoch.

fn random_u64(&mut self) -> u64

Returns a random number.

This number is the same in all replicas.

fn random_byte_array(&mut self) -> [u8; 32]

Returns a random byte array with 32 bytes.

This number is the same in all replicas.

fn execute_nns_function(
&self,
proposal_id: u64,
update: &ExecuteNnsFunction
) -> Result<(), GovernanceError>

Execute a ExecuteNnsFunction. The standard implementation is expected to call out to another canister and eventually report the result back

Loading content...