Module ic_nns_governance::pb::v1::manage_neuron

Nested message and enum types in ManageNeuron.

Modules

configure
Nested message and enum types in Configure.

disburse
Nested message and enum types in Disburse.

Structs

AddHotKey
Add a new hot key that can be used to manage the neuron. This provides an alternative to using the controller principal’s cold key to manage the neuron, which might be onerous and difficult to keep secure, especially if it is used regularly. A hot key might be a WebAuthn key that is maintained inside a user device, such as a smartphone.

Configure
Commands that only configure a given neuron, but do not interact with the outside world. They all require the caller to be the controller of the neuron.

Disburse
Disburse this neuron's stake: transfer the staked ICP to the specified account.

DisburseToNeuron
Disburse a portion of this neuron's stake into another neuron. This allows to split a neuron but with a new dissolve delay and owned by someone else.

Follow
Add a rule that enables the neuron to vote automatically on proposals that belong to a specific topic, by specifying a group of followee neurons whose majority vote is followed. The configuration of such follow rules can be used to a) distribute control over voting power amongst multiple entities, b) have a neuron vote automatically when its owner lacks time to evaluate newly submitted proposals, c) have a neuron vote automatically when its own lacks the expertise to evaluate newly submitted proposals, and d) for other purposes. A follow rule specifies a set of followees. Once a majority of the followees votes to adopt or reject a proposal belonging to the specified topic, the neuron votes the same way. If it becomes impossible for a majority of the followees to adopt (for example, because they are split 50-50 between adopt and reject), then the neuron votes to reject. If a rule is specified where the proposal topic is UNSPECIFIED, then it becomes a catch-all follow rule, which will be used to vote automatically on proposals belonging to topics for which no specific rule has been specified.

IncreaseDissolveDelay
The dissolve delay of a neuron can be increased up to a maximum of 8 years.

RegisterVote
Have the neuron vote to either adopt or reject a proposal with a specified id.

RemoveHotKey
Remove a hot key that has been previously assigned to the neuron.

Spawn
When the maturity of a neuron has risen above a threshold, it can be instructed to spawn a new neuron. This creates a new neuron that locks a new balance of ICP on the ledger. The new neuron can remain controlled by the same principal as its parent, or be assigned to a new principal.

Split
Split this neuron into two neurons.

StartDissolving

StopDissolving

Enums

Command