Struct ic_nns_governance::pb::v1::Neuron

pub struct Neuron {
    pub id: Option<NeuronId>,
    pub account: Vec<u8>,
    pub controller: Option<PrincipalId>,
    pub hot_keys: Vec<PrincipalId>,
    pub cached_neuron_stake_e8s: u64,
    pub neuron_fees_e8s: u64,
    pub created_timestamp_seconds: u64,
    pub aging_since_timestamp_seconds: u64,
    pub followees: HashMap<i32, Followees>,
    pub recent_ballots: Vec<BallotInfo>,
    pub kyc_verified: bool,
    pub transfer: Option<NeuronStakeTransfer>,
    pub maturity_e8s_equivalent: u64,
    pub not_for_profit: bool,
    pub dissolve_state: Option<DissolveState>,
}

This structure represents a neuron "at rest" in governance system of the Internet Computer IC.

Fields
id: Option<NeuronId>

The id of the neuron.

This is stored here temporarily, since its also stored on the map that contains neurons.

Initialization uses ids for the following graph. We need neurons to come into existence at genesis with pre-chosen ids, so a neuron needs to have an id. We could alternatively choose a unique naming scheme instead and chose the ids on the initialization of the canister.

account: Vec<u8>

The principal of the ICP ledger account where the locked ICP balance resides. This principal is indistinguishable from one identifying a public key pair, such that those browsing the ICP ledger cannot tell which balances belong to neurons.

controller: Option<PrincipalId>

The principal that actually controls the neuron. The principal must identify a public key pair, which acts as a “master key”, such that the corresponding secret key should be kept very secure. The principal may control many neurons.

hot_keys: Vec<PrincipalId>

Keys that can be used to perform actions with limited privileges without exposing the secret key corresponding to the principal e.g. could be a WebAuthn key.

cached_neuron_stake_e8s: u64

The amount of staked ICP tokens, measured in fractions of 10E-8 of an ICP.

Cached record of the locked ICP balance on the ICP ledger.

For neuron creation: has to contain some minimum amount. A spawned neuron with less stake cannot increase its dissolve delay.

neuron_fees_e8s: u64

The amount of ICP that this neuron has forfeited due to making proposals that were subsequently rejected or from using the 'manage neurons through proposals' functionality. Must be smaller than 'neuron_stake_e8s'. When a neuron is disbursed, these ICP will be burned.

created_timestamp_seconds: u64

When the Neuron was created. A neuron can only vote on proposals submitted after its creation date.

aging_since_timestamp_seconds: u64

The timestamp, in seconds from the Unix epoch, corresponding to the time this neuron has started aging. This is either the creation time or the last time at which the neuron has stopped dissolving.

This value is meaningless when the neuron is dissolving, since a dissolving neurons always has age zero. The canonical value of this field for a dissolving neuron is u64::MAX.

followees: HashMap<i32, Followees>

Map Topic to followees. The key is represented by an integer as Protobuf does not support enum keys in maps.

recent_ballots: Vec<BallotInfo>

Information about how this neuron voted in the recent past. It only contains proposals that the neuron voted yes or no on.

kyc_verified: bool

true if this neuron has passed KYC, false otherwise

transfer: Option<NeuronStakeTransfer>

The record of the transfer that was made to create this neuron.

maturity_e8s_equivalent: u64

The accumulated maturity of the neuron, in "e8s equivalent".

The unit is "e8s equivalent" to insist that, while this quantity is on the same scale as ICPs, maturity is not directly convertible to ICPs: conversion requires a minting event to spawn a new neuron.

not_for_profit: bool

Whether this neuron is "Not for profit", making it dissolvable by voting.

dissolve_state: Option<DissolveState>

At any time, at most one of when_dissolved and dissolve_delay are specified.

Locked. This is represented by dissolve_delay being set to a non zero value.

Dissolving. This is represented by when_dissolved being set, and this value is in the future.

Dissolved. All other states represent the dissolved state. That is, (a) when_dissolved is set and in the past, (b) dissolve_delay is set to zero, (c) neither value is set.

Cf. Neuron::stop_dissolving and Neuron::start_dissolving.

Implementations

impl Neuron

pub fn dissolve_delay_seconds(&self, now_seconds: u64) -> u64

Returns the dissolve delay of this neuron. For a locked neuron, this is just the recorded dissolve delay; for a dissolving neuron, this is the the time left (from now_seconds) until the neuron becomes dissolved; for a dissolved neuron, this function returns zero.

pub fn configure(
&mut self,
caller: &PrincipalId,
now_seconds: u64,
cmd: &Configure
) -> Result<(), GovernanceError>

Apply the specified neuron configuration operation on this neuron.

See manage_neuron::Configure for details.

pub fn get_neuron_info(&self, now_seconds: u64) -> NeuronInfo

Get the 'public' information associated with this neuron.

pub fn stake_e8s(&self) -> u64

Return the current 'stake' of this Neuron in number of 10^-8 ICPs. (That is, if the stake is 1 ICP, this function will return 10^8).

The stake can be decreased by making proposals that are subsequently rejected, and increased by transferring funds to the acccount of this neuron and then refreshing the stake.