Struct ic_nns_governance::stable_mem_utils::BufferedStableMemWriter

pub struct BufferedStableMemWriter { /* fields omitted */ }

An implementation of BufMut that writes to stable memory in chunks. Chunk size is specified on initialization, in BufferedStableMemWriter::new.

Note that you need to drop this, or call flush(), after using the BufMut methods, to write the buffer contents into the stable memory.

Implementations

impl BufferedStableMemWriter

pub fn new(buffer_size_bytes: u32) -> Self

Create a buffered writer with the given buffer size.

pub fn flush(&mut self)

Write the buffer contents to stable memory.